move logic around
